refactor: 메서드명 변경 및 이동
